ai/pyproject.toml
[project]
name = "sokoleo-langgraph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=7.2.0",
    "langchain>=1.2.12",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.2",
    "langchain-tavily>=0.2.17",
    "langgraph>=1.1.2",
    "langsmith>=0.7.17",
    "python-dotenv>=1.2.2",
]
